Modify DeadTreesElimination to support rdbars
A treetop hanging a read barrier can be removed if there is no other
real use of the read barrier. Otherwise the treetop has to be kept to
anchor the read barrier.

This change contains 2 parts:

The first part changes trivial DeadTreesElimination to remove the
first treetop from trees in the following form:
treetop
   rdbar
anchor
   =>rdbar

The second part prevents rdbar hanging under treetop with referenceCount
> 1 from being removed.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>